requirements.txt
flask
psycopg2-binary
sklearn
sqlparse
requests
